Move P2 customers to standalone page
